Sorting Changes

+++ StudentManagementSystem/Controllers/CollegesController.cs
@@ -14,40 +14,66 @@ namespace StudentManagementSystem.Controllers
     {
         private StudentInfoDbContext db = new StudentInfoDbContext();
 
+        public CollegesController()
+        {
+            db.Configuration.ValidateOnSaveEnabled = false;
+        }
+
         // GET: Colleges
         public ActionResult Index()
         {
-            return View(db.Colleges.ToList());
+            try
+            {
+                return View(db.Colleges.ToList());
+            }
+            catch(Exception ex)
+            {
+                return View("ErrorPage");
+            }
         }
 
         // GET: Colleges/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            try
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                College college = db.Colleges.Find(id);
+                if (college == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(college);
             }
-            College college = db.Colleges.Find(id);
-            if (college == null)
+            catch (Exception ex)
             {
-                return HttpNotFound();
+                return View("ErrorPage");
             }
-            return View(college);
         }
 
         // GET: Colleges/Create
         public ActionResult Create()
         {
-            College college = new College();
-            int i = 1;
-            List<Course> coursesCollection = db.Courses.ToList();
-            if (coursesCollection != null)
+            try
             {
-                coursesCollection.ForEach(
-                    item => college.CoursesList.Add(new CheckBoxModel() { Name = item.CourseName, ID = i++, Value = item.NoOfYears })
-                  );
+                College college = new College();
+                int i = 1;
+                List<Course> coursesCollection = db.Courses.ToList();
+                if (coursesCollection != null)
+                {
+                    coursesCollection.ForEach(
+                        item => college.CoursesList.Add(new CheckBoxModel() { Name = item.CourseName, ID = i++, Value = item.NoOfYears })
+                      );
+                }
+                return View(college);
             }
-            return View(college);
+            catch (Exception ex)
+            {
+                return View("ErrorPage");
+            }
         }
 
         // POST: Colleges/Create
@@ -57,44 +83,57 @@ namespace StudentManagementSystem.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "CollegeId,CollegeName,City,State,CoursesList")] College college)
         {
-            if (ModelState.IsValid)
+            try
             {
-                var selectedItem = college.CoursesList.Where(x => x.IsSelected);
-                foreach (var item in selectedItem)
+                if (ModelState.IsValid)
                 {
-                    var subject = db.Courses.FirstOrDefault(x => x.CourseName == item.Name);
-                    college.Courses.Add(subject);
-                    subject.Colleges.Add(college);
+                    var selectedItem = college.CoursesList.Where(x => x.IsSelected);
+                    foreach (var item in selectedItem)
+                    {
+                        var subject = db.Courses.FirstOrDefault(x => x.CourseName == item.Name);
+                        college.Courses.Add(subject);
+                        subject.Colleges.Add(college);
+                    }
+                    db.Colleges.Add(college);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
                 }
-                db.Colleges.Add(college);
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                return View(college);
             }
-
-            return View(college);
+            catch (Exception ex)
+            {
+                return View("ErrorPage");
+            }
         }
 
         // GET: Colleges/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            try
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            College college = db.Colleges.Find(id);
-            if (college == null)
-            {
-                return HttpNotFound();
-            }
-            int i = 0;
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                College college = db.Colleges.Find(id);
+                if (college == null)
+                {
+                    return HttpNotFound();
+                }
+                int i = 0;
 
-            List<Course> subjectCollection = new List<Course>(college.Courses);
-            foreach (var each in subjectCollection)
-            {
-                college.CoursesList.Add(new CheckBoxModel() { Name = each.CourseName, ID = i++, IsSelected = true, Value = each.NoOfYears });
+                List<Course> subjectCollection = new List<Course>(college.Courses);
+                foreach (var each in subjectCollection)
+                {
+                    college.CoursesList.Add(new CheckBoxModel() { Name = each.CourseName, ID = i++, IsSelected = true, Value = each.NoOfYears });
+                }
+                db.Courses.ToList().Except(college.Courses).ToList().ForEach(item => college.CoursesList.Add(new CheckBoxModel() { Name = item.CourseName, ID = i++, Value = item.NoOfYears }));
+                return View(college);
             }
-            db.Courses.ToList().Except(college.Courses).ToList().ForEach(item => college.CoursesList.Add(new CheckBoxModel() { Name = item.CourseName, ID = i++, Value = item.NoOfYears }));
-            return View(college);
+            catch (Exception ex)
+            {
+                return View("ErrorPage");
+            }
         }
 
         // POST: Colleges/Edit/5
@@ -104,41 +143,55 @@ namespace StudentManagementSystem.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "CollegeId,CollegeName,City,State,CoursesList")] College college)
         {
-            if (ModelState.IsValid)
+            try
             {
-                var exisitingSemesters = db.Colleges.Where(s => s.CollegeId == college.CollegeId).Include(y => y.Courses);
-                var cour = exisitingSemesters.FirstOrDefault();
-                cour.CollegeName = college.CollegeName;
-                cour.City = college.City;
-                cour.State = college.State;
-
-                cour.Courses.Clear();
-                var selectedItem = college.CoursesList.Where(x => x.IsSelected);
-                foreach (var item in selectedItem)
+                if (ModelState.IsValid)
                 {
-                    var course = db.Courses.FirstOrDefault(x => x.CourseName == item.Name);
-                    cour.Courses.Add(course);
+                    var exisitingSemesters = db.Colleges.Where(s => s.CollegeId == college.CollegeId).Include(y => y.Courses);
+                    var cour = exisitingSemesters.FirstOrDefault();
+                    cour.CollegeName = college.CollegeName;
+                    cour.City = college.City;
+                    cour.State = college.State;
+
+                    cour.Courses.Clear();
+                    var selectedItem = college.CoursesList.Where(x => x.IsSelected);
+                    foreach (var item in selectedItem)
+                    {
+                        var course = db.Courses.FirstOrDefault(x => x.CourseName == item.Name);
+                        cour.Courses.Add(course);
+                    }
+                    db.Entry(cour).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
                 }
-                db.Entry(cour).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                return View(college);
             }
-            return View(college);
+            catch (Exception ex)
+            {
+                return View("ErrorPage");
+            }
         }
 
         // GET: Colleges/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            try
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                College college = db.Colleges.Find(id);
+                if (college == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(college);
             }
-            College college = db.Colleges.Find(id);
-            if (college == null)
+            catch (Exception ex)
             {
-                return HttpNotFound();
+                return View("ErrorPage");
             }
-            return View(college);
         }
 
         // POST: Colleges/Delete/5
@@ -146,10 +199,23 @@ namespace StudentManagementSystem.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            College college = db.Colleges.Find(id);
-            db.Colleges.Remove(college);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            try
+            {
+                College college = db.Colleges.Find(id);
+                if (college.Students != null && college.Students.Count > 0)
+                {
+                    ViewBag.StudentNames = college.Students.Select(x => x.StudentName).ToList();
+                    ViewBag.Delete = "College";
+                    return View("Error");
+                }
+                db.Colleges.Remove(college);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            catch (Exception ex)
+            {
+                return View("ErrorPage");
+            }
         }
 
         protected override void Dispose(bool disposing)
